Start port of Circus Linux
This is the start of a port to Android of the game
Circus Linux ( http://www.newbreedsoftware.com/circus-linux ).

It utilizes the framework of Pelya's port of the SDL library to Android.
You can read more about this in the file pelya.txt.

The base directory of circuslinux is project/jni/application/circuslinux

We have pulled down the source code of circuslinux (from Ubuntu) and put
it in the circuslinux subdirectory of the base circuslinux directory.

We recursively zipped up the data directory and moved it to the
AndroidData subdirectory of the base circuslinux directory.  We then
split the zip file into one meg chunks.

We pulled an image from the provided images, and placed it at icon.png

We also added a AndroidAppSettings.cfg file.