More keyboard changes
We pop the on-screen keyboard up when high score has been achieved.
I modified the hint in EditText to say press Enter, not Back.

One thing not dealt with is where to put the edittext.  Right now it is
a magic number kludge that only works for Samsung Galaxy Tab 10.1
devices.  This has to be generalized more.