Turn KEYPAD functionality on
Turn the KEYPAD functionality on.

I should be able to turn this on in globals.h and have it work, but
that is not happening, so I modified game.c and options.c for now
to always default to KEYPAD on.  This can be looked at again in
the future.