Modify various ifdef's for Android
The upstream source, ri-li, has various ifdefs in utils.cpp which
set different parameters depending upon if the target is WINDOWS
(Windows) or MACLINUX (Mac or Linux).  For the definitions given,
sometimes Android needs the MACLINUX definition, sometimes it needs
the WINDOWS definition.

We should sort all of this out properly so that it can target all
systems and integrate nicely with upstream, but I just did a little
kludge to where it uses the proper definitions and only targets
Android.  We can clean this up in the future.